Move flow fixtures into test_flow/conftest.py
None of our fixtures are used outside of the flow tests, so we can move
them into a separate conftest.py in the test_flow directory. This has
the advantage of separating our test fixtures from our test hooks
(command line options and marker manipulation).
